# public/docs/v1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is MyMusic API,a demo for exploring API design best practices in Rails 5.\nTo be able to try operations you have to obtain a JSON Web Token, fill  in the token field and press the arrow button next to it.\n",
        "version": "1.0.0",
        "title": "MyMusic",
        "termsOfService": "",
        "contact": {
            "name": "Spiros Kabasakalis",
            "email": "kabasakalis@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "host": "api.app.me:3000",
    "basePath": "/v1",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/vnd.siren+json"
    ],
    "paths": {
...
        "/users": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Finds all users",
                "description": "Gets all users,paginated result.",
                "operationId": "getAllUsers",
                "parameters": [
                    {
                        "in": "query",
                        "name": "role",
                        "description": "Filter by role",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "user",
                            "admin",
                            "superadmin"
                        ]
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Filter by ids",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/User"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Add a new user",
                "description": "Only Admins and SuperAdmins can create users",
                "operationId": "addUser",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "User to add.You can also create playlists for this user if you are authorized.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/User"
                        }
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/genres": {
            "get": {
                "tags": [
                    "genre"
                ],
                "summary": "Finds all genres",
                "description": "Gets all genres,paginated result.",
                "operationId": "getAllGenres",
                "parameters": [
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Filter by ids",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Genre"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "genre"
                ],
                "summary": "Add a new genre",
                "description": "Only Admins and SuperAdmins can create genres",
                "operationId": "addGenre",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Genre to add",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Genre"
                        }
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/artists": {
            "get": {
                "tags": [
                    "artist"
                ],
                "summary": "Finds all Artists",
                "description": "Gets all artists,paginated result.",
                "operationId": "getAllArtists",
                "parameters": [
                    {
                        "in": "query",
                        "name": "country",
                        "description": "Filter by country",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "genre",
                        "description": "Filter by genre",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Filter by ids",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Artist"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "artist"
                ],
                "summary": "Add a new artist",
                "description": "Only Admins and SuperAdmins can create Artists",
                "operationId": "addArtist",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Artist to add.You can either provide a genre_id (list Genres to find the id) to associate the artist with an existing Genre or provide the title of a new Genre,(in which case you leave genre_id blank).You can also create and associate albums for the new Artist.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Artist"
                        }
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/albums": {
            "get": {
                "tags": [
                    "album"
                ],
                "summary": "Finds all albums",
                "description": "Gets all albums,paginated result.",
                "operationId": "getAllAlbums",
                "parameters": [
                    {
                        "in": "query",
                        "name": "year",
                        "description": "Filter by Year",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "artist",
                        "description": "Filter by artist",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "in": "query",
                        "name": "id",
                        "description": "Filter by ids",
                        "required": false,
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi"
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Album"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "album"
                ],
                "summary": "Add a new album",
                "description": "Only Admins and SuperAdmins can create albums",
                "operationId": "addAlbum",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Album to add.You can provide an artist_id (list artists to find the id) to associate the album with an existing artist. You can also create and associate tracks for the new album.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Album"
                        }
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/playlists": {
            "get": {
                "tags": [
                    "playlist"
                ],
                "summary": "Finds all playlists",
                "description": "Gets all playlists,paginated result.",
                "operationId": "getAllPlaylists",
                "parameters": [
                    {
                        "in": "query",
                        "name": "user",
                        "description": "Filter by User's ID",
                        "required": false,
                        "type": "string"
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Playlist"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            },
            "post": {
                "tags": [
                    "playlist"
                ],
                "summary": "Add a new playlist",
                "description": "Only Admins and SuperAdmins can create playlists",
                "operationId": "addPlaylist",
                "parameters": [
                    {
                        "in": "body",
                        "name": "body",
                        "description": "Playlist to add.You can provide a user_id (list users  to find the id) to associate the playlist with an existing user. You can also create and associate albums and tracks for the new playlist.",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Playlist"
                        }
                    }
                ],
                "security": [
                    {
                        "Authorization": []
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Authorization": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "User": {
            "type": "object",
            "properties": {
                "user": {
                    "type": "object",
                    "required": [
                        "email",
                        "password",
                        "role"
                    ],
                    "properties": {
                        "name": {
                            "type": "string",
                            "description": "username"
                        },
                        "email": {
                            "type": "string",
                            "description": "email"
                        },
                        "password": {
                            "type": "string",
                            "description": "password (at least 8 characters)"
                        },
                        "password_confirmation": {
                            "type": "string",
                            "description": "password confirmation"
                        },
                        "role": {
                            "type": "string",
                            "enum": [
                                "user",
                                "admin",
                                "superadmin"
                            ]
                        },
                        "playlists_attributes": {
                            "description": "playlists that this track belongs to",
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "Genre": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "description": "Name of new Genre to add."
                }
            }
        },
        "Artist": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "object",
                    "required": [
                        "title"
                    ],
                    "properties": {
                        "genre_id": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "title": {
                            "type": "string",
                            "example": "Myrath",
                            "description": "Name of band or album."
                        },
                        "country": {
                            "type": "string",
                            "example": "Tunisia",
                            "description": "artist's country of origin."
                        },
                        "genre_attributes": {
                            "$ref": "#/definitions/Genre"
                        },
                        "albums_attributes": {
                            "description": "artist's albums",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AlbumNested"
                            }
                        }
                    }
                }
            }
        },
...
        "Album": {
            "type": "object",
            "properties": {
                "album": {
                    "type": "object",
                    "required": [
                        "title"
                    ],
                    "properties": {
                        "artist_id": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "title": {
                            "type": "string"
                        },
                        "year": {
                            "type": "integer",
                            "format": "int32"
                        },
                        "tracks_attributes": {
                            "description": "Tracks on this album",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TrackNested"
                            }
                        }
                    }
                }
            }
        },
        "AlbumNested": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string"
                },
                "year": {
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
...
        "TrackNested": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "Playlist": {
            "type": "object",
            "properties": {
                "playlist": {
                    "type": "object",
                    "required": [
                        "title"
                    ],
                    "properties": {
                        "title": {
                            "type": "string"
                        },
                        "user_id": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "tracks_attributes": {
                            "description": "tracks in this playlist",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/TrackNested"
                            }
                        },
                        "albums_attributes": {
                            "description": "albums in this playlist",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/AlbumNested"
                            }
                        }
                    }
                }
            }
        },
...
    }
}